[refactor]: dto 대신 LadderGame에서 LadderMap 생성하도록 변경
